ディレクトリ名変更によるimportパスの変更
